Autogenerate more files
